Fixed code review issues
